+++ pymongo/mongo_client.py
@@ -507,15 +507,15 @@ class MongoClient(common.BaseObject):
     def max_pool_size(self):
         """The maximum number of sockets the pool will open concurrently.
 
+        When the pool has reached `max_pool_size`, operations block waiting for
+        a socket to be returned to the pool. If ``waitQueueTimeoutMS`` is set,
+        a blocked operation will raise :exc:`~pymongo.errors.ConnectionFailure`
+        after a timeout. By default ``waitQueueTimeoutMS`` is not set.
+
         .. warning:: SIGNIFICANT BEHAVIOR CHANGE in 2.5.1+. Previously, this
-          parameter would limit only the idle connections the pool would hold
+          parameter would limit only the idle sockets the pool would hold
           onto, not the number of open sockets. The default has also changed
           to 100.
-
-        .. note:: ``max_pool_size`` caps the number of concurrent
-          connections to the server. Connection or query attempts when the pool
-          has reached `max_pool_size` will block until conn_timeout or a
-          connection has been returned to the pool.
 
         .. versionchanged:: 2.5.1+
         .. versionadded:: 1.11